Refactored minMax table functionality out to table.ts
